+++ packages/core/src/core-providers/structure/structure.state.ts
@@ -1,38 +1,11 @@
 import { declareAtom } from '@reatom/core';
 import noop from '@tinkoff/utils/function/noop';
-import toPairs from '@tinkoff/utils/object/toPairs';
 import { map, shareReplay } from 'rxjs/operators';
 import { toRxStore } from '../../base/store';
-
-export const toFieldsObj = (obj: FormStructure) => {
-  const flattened = {};
-
-  const extractFields = (str: FormStructure) =>
-    toPairs(str).forEach(([key, value]) => {
-      if ('group' in value) {
-        return extractFields(value.group);
-      }
-      flattened[key] = value;
-    });
-
-  extractFields(obj);
-  return flattened;
-};
+import { toFieldsObj, getInitialState } from './structure.util';
 
 import type { Store } from '@reatom/core';
 import type { FormStructure, StructureFactory } from './structure.types';
-
-export type State = FormStructure;
-
-const getInitialState = (factory: StructureFactory) => {
-  return factory({
-    group: (a, opts) => ({
-      type: 'group',
-      group: a,
-      opts: opts || {},
-    }),
-  });
-};
 
 export const useState = ({
   globalStore,
@@ -41,15 +14,19 @@ export const useState = ({
   globalStore: Store;
   factory: StructureFactory;
 }) => {
-  const initialState = getInitialState(factory);
-  const atom = declareAtom<State>(['structure'], initialState, () => []);
+  const initialConfig = getInitialState(factory);
+  const atom = declareAtom<FormStructure>(
+    ['structure'],
+    initialConfig,
+    () => [],
+  );
 
   globalStore.subscribe(atom, noop);
 
   return {
     _atom: atom,
     rx: toRxStore(globalStore, atom).pipe(map(toFieldsObj), shareReplay()),
-    initial: toFieldsObj(initialState),
-    initialConfig: initialState,
+    initialState: toFieldsObj(initialConfig),
+    initialConfig: initialConfig,
   };
 };
